+++ renderer/components/ConnectionsSection/index.ts
@@ -1,0 +1,1 @@
+export { default as ConnectionsSection } from './ConnectionsSection';
