+++ renderer/components/ReportsFolderSection/index.ts
@@ -1,0 +1,1 @@
+export { default as ReportsFolderSection } from './ReportsFolderSection';
