+++ renderer/components/ThemeSection/index.ts
@@ -1,0 +1,1 @@
+export { default as ThemeSection } from './ThemeSection';
