+++ renderer/components/VersionSection/index.ts
@@ -1,0 +1,1 @@
+export { default as VersionSection } from './VersionSection';
